requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
Authlib==1.4.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cramjam==2.9.1
cryptography==44.0.1
distro==1.9.0
Django==3.1.12
django-cors-headers==3.1.1
djangorestframework==3.11.1
djongo==1.3.7
dnspython==2.7.0
fastapi==0.115.8
ffmpeg==1.4
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
grpclib==0.4.7
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
modal==0.73.64
multidict==6.1.0
openai==1.64.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pykerberos==1.2.4
pymongo==3.11.4
PyMuPDF==1.25.3
python-dotenv==1.0.1
python-snappy==0.7.3
pytz==2025.1
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
shellingham==1.5.4
sigtools==4.0.1
sniffio==1.3.1
sqlparse==0.2.4
starlette==0.45.3
synchronicity==0.9.11
toml==0.10.2
tqdm==4.67.1
typer==0.15.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.12.2
urllib3==2.3.0
watchfiles==1.0.4
yarl==1.18.3
